requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
pypdf>=4.0.0
cryptography>=3.1
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
